% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_costs_by_work_package.R
\name{append_zero_costs}
\alias{append_zero_costs}
\title{Helper function: append zero costs}
\usage{
append_zero_costs(x, n_work_packages)
}
\arguments{
\item{x}{x}

\item{n_work_packages}{number of work packages in EXCEL template
(default: 7, as used for DWC)}
}
\value{
data frame with zero costs
}
\description{
Helper function: append zero costs
}
