% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partner_template_xls.R
\name{create_partner_template_xls}
\alias{create_partner_template_xls}
\title{Create an Excel File for Project Partner Information}
\usage{
create_partner_template_xls(
  n_partners = 7L,
  file_name = "project_partners_tmp.xlsx",
  sheet_name = "Partners-PIC-Main contact"
)
}
\arguments{
\item{n_partners}{number of partners}

\item{file_name}{name of Excel file. Default: "project_partners_tmp.xlsx".}

\item{sheet_name}{name of Excel sheet. Default: "Partners-PIC-Main contact".}
}
\description{
Create an Excel File for Project Partner Information
}
