% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_matching_files.R
\name{download_matching_files}
\alias{download_matching_files}
\title{Download Files from Nextcloud that Match a File Name Pattern}
\usage{
download_matching_files(path, pattern)
}
\arguments{
\item{path}{Nextcloud path}

\item{pattern}{file name pattern}
}
\description{
Download Files from Nextcloud that Match a File Name Pattern
}
