% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_costs_by_type.R
\name{get_costs_by_sector}
\alias{get_costs_by_sector}
\title{Helper function: prepare costs by sector}
\usage{
get_costs_by_sector(costs_overview)
}
\arguments{
\item{costs_overview}{costs data}
}
\value{
df with costs by type (e.g. health, water, it, ..)
}
\description{
Helper function: prepare costs by sector
}
