% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cost_data_by_type.R
\name{get_costs_by_type}
\alias{get_costs_by_type}
\title{Helper function: prepare costs by type}
\usage{
get_costs_by_type(costs_overview)
}
\arguments{
\item{costs_overview}{costs data}
}
\value{
df with costs by type (e.g. SME, research, ..)
}
\description{
Helper function: prepare costs by type
}
