% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_costs_by_work_package.R
\name{get_costs_by_work_package}
\alias{get_costs_by_work_package}
\title{Get Costs by Work Package}
\usage{
get_costs_by_work_package(costs)
}
\arguments{
\item{costs}{list with "costs" from multiple partner Excelsheets}
}
\value{
data frame with costs per work package
}
\description{
Get Costs by Work Package
}
