% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_partner_budget_from_excel.R
\name{get_named_excel_ranges}
\alias{get_named_excel_ranges}
\title{Helper function: get named Excel ranges}
\usage{
get_named_excel_ranges(file)
}
\arguments{
\item{file}{full path to EXCEL file}
}
\value{
named Excel ranges
}
\description{
Helper function: get named Excel ranges
}
