% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_partner_budget_versions.R
\name{list_partner_budget_versions}
\alias{list_partner_budget_versions}
\title{List the Partner's File Versions (Except Current)}
\usage{
list_partner_budget_versions(
  path = "proposals/h2020_covid/60_Budget/10_Filled_out_forms"
)
}
\arguments{
\item{path}{path to folder containing Excel files on Nextcloud}
}
\description{
List the Partner's File Versions (Except Current)
}
