% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_or_write_csv}
\alias{read_or_write_csv}
\title{Read CSV File or Write CSV File if File does not exist}
\usage{
read_or_write_csv(file, data = NULL)
}
\arguments{
\item{file}{path to CSV file}

\item{data}{optional data to be written to CSV file in case that file does
not exist}
}
\description{
Read CSV File or Write CSV File if File does not exist
}
