% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_partners_budget_from_excel.R
\name{read_partners_budget_from_excel}
\alias{read_partners_budget_from_excel}
\title{Read Partners Budget From Excel Files}
\usage{
read_partners_budget_from_excel(
  files,
  n_work_packages = 7,
  run_parallel = FALSE
)
}
\arguments{
\item{files}{full path to EXCEL files}

\item{n_work_packages}{number of work packages in EXCEL template
(default: 7, as used for DWC)}

\item{run_parallel}{should import be performed using multiple CPU cores or
only run on a single core (default: FALSE)}
}
\value{
list with imported EXCEL budget files data
}
\description{
Read Partners Budget From Excel Files
}
