% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cost_matrices.R
\name{to_cost_matrices}
\alias{to_cost_matrices}
\title{Summarise all Cost Tables}
\usage{
to_cost_matrices(costs_by_wp)
}
\arguments{
\item{costs_by_wp}{list of cost tables, as returned by
\code{kwb.budget:::get_all_cost_sheets} in element \code{by_wp}}
}
\description{
Summarise all Cost Tables
}
