% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upload_files}
\alias{upload_files}
\title{Upload Files to Nextcloud}
\usage{
upload_files(files, target_path)
}
\arguments{
\item{files}{path(s) to local file(s)}

\item{target_path}{Nextcloud target path}
}
\description{
Upload Files to Nextcloud
}
