% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_costs_by_work_package.R
\name{append_zero_costs}
\alias{append_zero_costs}
\title{Helper function: append zero costs}
\usage{
append_zero_costs(x)
}
\arguments{
\item{x}{x}
}
\value{
data frame with zero costs
}
\description{
Helper function: append zero costs
}
