% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partners_budget_files.R
\name{create_partners_budget_files}
\alias{create_partners_budget_files}
\title{Create partners budget files (for DWH)}
\usage{
create_partners_budget_files(
  partner_info,
  path_budget_template,
  prefix = "",
  target_dir = tempdir(),
  set_values = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{partner_info}{data frame as returned by \code{\link{read_partner_info}}}

\item{path_budget_template}{path to budget template EXCEL file}

\item{prefix}{prefix given to all created files (default: "")}

\item{target_dir}{target directory where to save the budget files (default:
file.path(dirname(path_budget_template), "10_Filled_out_forms"))}

\item{set_values}{should metadata from partners EXCEL file be set or just the
template budget EXCEL file copied}

\item{overwrite}{should existing EXCEL files be overwritten (default: TRUE)}
}
\value{
paths to created Excel files
}
\description{
Create partners budget files (for DWH)
}
