% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partners_budget_files.R
\name{create_partners_budget_files}
\alias{create_partners_budget_files}
\title{Create partners budget files (for DWH)}
\usage{
create_partners_budget_files(
  path_partners,
  path_budget_template,
  project_shortname = "DWH",
  target_dir = file.path(dirname(path_budget_template), "10_Filled_out_forms"),
  set_values = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{path_partners}{path to partners budget metadata EXCEL file}

\item{path_budget_template}{path to budget template EXCEL file}

\item{project_shortname}{proposal acronym (default: "DWH")}

\item{target_dir}{target directory where to save the budget files (default:
file.path(dirname(path_budget_template), "10_Filled_out_forms"))}

\item{set_values}{should metadata from partners EXCEL file be set or just the
template budget EXCEL file copied}

\item{overwrite}{should existing EXCEL files be overwritten (default: TRUE)}
}
\value{
budget excel files for each project partner
}
\description{
Create partners budget files (for DWH)
}
