% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_partner_info.R
\name{read_partner_info}
\alias{read_partner_info}
\title{Read Information on Project Partners from (Downloaded) Excel file}
\usage{
read_partner_info(nextcloud_path, sheet, local_path = NULL, columns = NULL)
}
\arguments{
\item{nextcloud_path}{path to Excel file on nextcloud}

\item{sheet}{name of Excel sheet}

\item{local_path}{if given, the Excel file is read from this (local) path}

\item{columns}{vector of column names. If given, only these columns are
returned}
}
\value{
data frame with attribute "local_path" being set to the path to the
local Excel file from which the data were read
}
\description{
Either nextcloud_path or local_path must be given with local_path taking
precedence over nextcloud_path.
}
